优化招标人管理页面中发布项目的项目预算调整可以有小数显示处理

+++ src/main/java/com/ztbdz/tenderee/pojo/Project.java
@@ -57,7 +57,7 @@ public class Project extends Model<Project> implements Serializable {
     @ApiModelProperty(value = "上传采购文件")
     private String procurementDocuments;
     @ApiModelProperty(value = "预算")
-    private Long money;
+    private Double money;
     @ApiModelProperty(value = "预算金额大写")
     @TableField(exist = false)
     private String moneyUppercase;
